requirements.txt
Django==5.2.4
djangorestframework==3.15.2
gunicorn==21.2.0
whitenoise==6.7.0
